+++ DoConnect-Backend/src/main/java/cogent/university/com/DoConnectBackend/repository/QuestionRepository.java
@@ -10,8 +10,8 @@ public interface QuestionRepository extends JpaRepository<Question, Long> {
 
 	List<Question> findBytopic(String topic);
 	List<Question> findBystatus(String status);
-	Question findById(int id);
-	void deleteById(int id);
+	Question findById(long id);
+	void deleteById(long id);
 
 
 	List<Question> findAllByTitle(String title);
